requirements.txt
aenum==2.2.4
aiogremlin==3.2.4
aiohttp==1.3.3
argon2-cffi==20.1.0
async-timeout==2.0.1
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
blinker==1.3
cassandra-driver==3.24.0
cffi==1.14.3
chardet==2.3.0
click==7.1.2
cloud-init==20.3
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
geomet==0.2.1.post1
gremlinpython==3.4.6
idna==2.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==5.3.0
ipython-genutils==0.2.0
ipython-gremlin==1.0.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
language-selector==0.1
MarkupSafe==0.23
matplotlib==3.0.3
mistune==0.8.4
multidict==5.0.0
nbconvert==5.6.1
nbformat==5.0.8
networkx==2.4
notebook==6.1.4
numpy==1.18.5
oauthlib==1.0.3
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==1.0.18
ptyprocess==0.6.0
pyasn1==0.1.9
pycparser==2.20
pycurl==7.43.0
pydot==1.4.1
Pygments==2.7.2
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.9
python-dateutil==2.8.1
python-debian==0.1.27
python-systemd==231
pytz==2020.1
PyYAML==3.12
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.9.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
ssh-import-id==5.5
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==0.9.8
zipp==1.2.0
